правки 3
