+++ bootstrap.php
@@ -1,4 +1,5 @@
 <?php
 require_once 'constants.php';
-require_once 'functions.php';
+require_once 'functions/functions.php';
+require_once 'functions/validate-functions.php';
 require_once 'sql-connect.php';
