+++ templates/add-posts/form-post.php
@@ -1,6 +1,15 @@
-<?php $addPost = $types[array_search($currentType, array_column($types, 'title'))]; ?>
-<section class="adding-post__<?= $addPost['title'] ?> tabs__content tabs__content--active">
-    <h2 class="visually-hidden">Форма добавления <?= isRusNameTypes($addPost['title']) ?></h2>
+<?php
+
+/**
+ * @var array $errors --массив ошибок валидации формы
+ * @var string $currentType -- текущий тип поста
+ * @var string $addPostContent -- вывод html страницы полей текущего поста
+ * @var array $errorTitleRus -- массив заголовков возможных ошибок
+ */
+
+?>
+<section class="adding-post__<?= $currentType ?> tabs__content tabs__content--active">
+    <h2 class="visually-hidden"><?= isRusNameTypes($currentType) ?></h2>
     <form class="adding-post__form form" action="/add.php?type=<?= $currentType ?>" method="post"
           enctype="multipart/form-data">
         <div class="form__text-inputs-wrapper">
@@ -8,13 +17,13 @@
                 <div class="adding-post__input-wrapper form__input-wrapper
                                     <?= isErrorCss($errors['heading']); ?>">
                     <label class="adding-post__label form__label" for="heading">Заголовок <span
-                                class="form__input-required">*</span></label>
+                            class="form__input-required">*</span></label>
                     <div class="form__input-section">
                         <input class="adding-post__input form__input" id="heading" type="text"
                                name="heading" placeholder="Введите заголовок"
                                value="<?= getPostVal('heading') ?>">
                         <button class="form__error-button button" type="button">!<span
-                                    class="visually-hidden">Информация об ошибке</span></button>
+                                class="visually-hidden">Информация об ошибке</span></button>
                         <div class="form__error-text">
                             <h3 class="form__error-title"><?= FORM_ERROR ?></h3>
                             <p class="form__error-desc"><?= $errors['heading'] ?></p>
@@ -30,7 +39,7 @@
                         <? foreach ($errors as $key => $value):
                             if (!empty($value)):?>
                                 <li class="form__invalid-item"><?= $errorTitleRus[$key] ?>. <?= $value ?>.</li>
-                            <? endif;
+                            <?endif;
                         endforeach; ?>
                     </ul>
                 </div>
@@ -46,8 +55,9 @@
                             <span>Перетащите фото сюда</span>
                         </div>
                     </div>
-                    <button class="adding-post__input-file-button form__input-file-button form__input-file-button--photo button"
-                            type="button">
+                    <button
+                        class="adding-post__input-file-button form__input-file-button form__input-file-button--photo button"
+                        type="button">
                         <span>Выбрать фото</span>
                         <svg class="adding-post__attach-icon form__attach-icon" width="10" height="20">
                             <use xlink:href="#icon-attach"></use>
